Fixes #209
Changed UDF from Standard Barcode to Standard File Barcode

For some reason Clarity does not let you change the types of UDFs.
Standard Barcode on Pico DTX was configured as a numeric which meant
that when it was saved, it was saved in eponential format. We use this
number to look up a file so really we need it as a string