Fixes #216
Added dilution factor to scale concentrations back up to stock plate. Obviously
this makes it a bit variable.

Fixed standard deviation of for intermediate coefficients. Was for some reason
dividing sum of square differences from mean by 3 when there are 4 control columns.

Still to look at whether Sample Managment want coefficient of variation looked at as
it is not calculating it correctly. However, this commit now being brings the Clarity
implementation to be equal to SequenceScape's.