Fixes #418 - Shadow stamping not creating correct output type
Moved the checking of options to a separate method, as there
were quite a few now. Added in a check to see if both shadow_plate
has been set and container_type_name has been provided. If both
true the user will get a warning and the container_type_name will
be clearer.

The container_type_name will then be determined from the input
container, which is what we want for shadow stamping.