Story C128 and more message queues...
Refacting of the message consumption. There is now a general message
consumer class that can be used to build multiple queues.

Also changed how messages are added to local message queue. Instead of
using the default routing key from config, one must be provided to the
messenger EPP.

Added a new report message consumer.

Added a new EPP to generate a 'manifest' report. Can be given a
process_url, a list of container ids, or a message object from
the message queue.